voice_pipeline/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
deepgram-sdk==3.2.7
google-genai
python-dotenv==1.0.1
httpx==0.27.0
starlette==0.37.2
